requirements.txt
numpy<2.3
pandas>=2.0.0
streamlit>=1.28.0
plotly>=5.18.0
yfinance>=0.2.28
pandas-ta>=0.3.14b0
google-genai>=1.0.0
python-dotenv>=1.0.0

kaleido>=0.2.1
